TURLİRA
To access the application, you use your browser to connect to a front end website written in SMS, CSS, and JS.

Then, instead of talking to a backend web server, this website talks directly to the blockchain. This is where all of the code and data and code for the application live.

This code is contained in Smart Contracts, written in with the Solidity Programming language, which looks a lot like JavaScript. These smart contracts are immutable, which means the the code cannot change.

All of the data is stored on the public ledger, which is also immutable. Any time we add new data to the blockchain, it will be permanent publicly verifiable.

So that’s an overview of what a blockchain is, and how it works.

Now let’s get a preview of the app that we’re going to build together and how it leverages the power for the blockchain.